requirements.txt
Django~=5.0.2
requests~=2.31.0
djangorestframework~=3.15.2
djangorestframework-simplejwt~=5.3.1